requirements.txt
discord.py
pynacl
aiohttp